+++ crates/plugin_partial_bundling/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3](https://github.com/farm-fe/farm/compare/farmfe_plugin_partial_bundling-v0.0.2...farmfe_plugin_partial_bundling-v0.0.3) - 2024-04-01
+
+### Other
+- updated the following local packages: farmfe_core, farmfe_testing_helpers
+
 ## [0.0.2](https://github.com/farm-fe/farm/compare/farmfe_plugin_partial_bundling-v0.0.1...farmfe_plugin_partial_bundling-v0.0.2) - 2024-03-24
 
 ### Other